refactor signin and register hooks
