+++ src/components/Auth/Login/SignInFormik.tsx
@@ -1,4 +1,4 @@
-import { Form, Formik } from 'formik';
+import { Form } from 'formik';
 import { FormField } from '@/components/UI/Forms/CustomInput/CustomInput';
 import { validationSchemaSignIn } from '../../../validation/validationSignIn';
 import s from './signIn.module.scss';
